refactor(decisions): extract the WWWD directional verdict ladder (BI-F5F2869D)
`evaluator.ts` crossed the 800-LOC ceiling at 818. The Module Size Guard treats
a newly-oversized file as a hard breach rather than something to re-baseline,
and it is right to: the file had accumulated two separate concerns.

The content-aware directional ladder — how a recorded stance becomes an
approve/decline verdict — is lifted whole into `directional-outcome.ts`. Pure
extraction: no branch reordered, no condition changed. The content-blind WWMD
path stays exactly where it was, and the existing suite is the proof the move
changed nothing (520 tests green across lib/decision-perspective,
lib/founder-review and lib/decision, unchanged from before the extraction).

evaluator.ts 818 -> 712 LOC; the new module is 144.

Plan updated with the re-scope and this extraction.

Design-Grounding-Decision: no-contract-change (pure extraction to satisfy the module-size ceiling; behaviour pinned by the existing suite)
